requirements.txt
Django
djangorestframework
markdown
django-filter
mysqlclient
django-environ
django-cors-headers
pyexcel_xls
pyexcel_xlsx
djangorestframework-simplejwt